+++ semana15/Express/package.json
@@ -3,8 +3,14 @@
     "cors": "^2.8.5",
     "express": "^4.17.1"
   },
+  "scripts": {
+		"start": "tsc && node ./build/index.js",
+    "dev-start": "tsnd src/index.ts",
+    "test": "echo \"Error: no test specified\" && exit 1"
+},
   "devDependencies": {
     "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13"
+    "@types/express": "^4.17.13",
+    "ts-node-dev": "^1.1.8"
   }
 }
