+++ semana15/Express/src/endpoint/getCountriesBySearch.ts
@@ -7,7 +7,35 @@ export const getCountriesBySearch = (
     res: Response
     
 ):void => {
-    const result: country | undefined = countries.find(
-        country => country.name === req.params.name
-    )
+    let countriesList: country[] = countries
+
+    if(req.query.name){
+        const query = req.query.name as string
+        countriesList = countriesList.filter(
+            country => country.name.toLocaleLowerCase().includes(query.toLocaleLowerCase())
+        )
+    }
+
+    if(req.query.capital){
+        const query = req.query.capital as string
+        countriesList = countriesList.filter(
+            country => country.capital.toLocaleLowerCase().includes(query.toLocaleLowerCase())
+        )
+    }
+
+    if(req.query.continent){
+        const query = req.query.continent as string
+        countriesList = countriesList.filter(
+            country => country.continent.toLocaleLowerCase().includes(query.toLocaleLowerCase())
+        )
+    }
+
+
+    if(countriesList.length > 0){
+        res.send(countriesList)
+    }else{
+        res.statusCode = 404
+        res.end(`Resultado não enconotrado`)
+    }
+
 }
